requirements.txt
# =====================================================
# Wiki Data Pipeline - Python Dependencies
# =====================================================

# Core Data Processing
pyspark==3.5.1
pandas==2.1.4
numpy==1.26.2

# Kafka
kafka-python==2.0.2

# HTTP Requests
requests==2.31.0
urllib3==2.1.0

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.23

# Configuration
python-dotenv==1.0.0

# Logging & Monitoring
prometheus-client==0.19.0

# Development
jupyter==1.0.0
ipython==8.18.1

# Testing
pytest==7.4.3
pytest-cov==4.1.0
